services/security/requirements.txt
# ──────────────────────────────────────────────
# Security & Privacy Service — runtime dependencies
# ──────────────────────────────────────────────
fastapi>=0.110,<1.0
mangum>=0.17,<1.0
pydantic>=2.6,<3.0
boto3>=1.34,<2.0
cryptography>=42.0,<44.0
python-jose[cryptography]>=3.3,<4.0
uvicorn>=0.29,<1.0
